+++ src/main/java/com/backend/api/repositories/CompanyRepository.java
@@ -11,9 +11,9 @@ import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository
-public interface CompanyRepository extends JpaRepository<Company, Integer>, JpaSpecificationExecutor<Company> {
+public interface CompanyRepository extends JpaRepository<Company, Long>, JpaSpecificationExecutor<Company> {
 
   @Transactional()
     @Query("SELECT DISTINCT c FROM User u JOIN u.companies c WHERE u.id = :userId")
-  List<Company> getMenus(@Param("userId") Integer userId);
+  List<Company> getMenus(@Param("userId") Long userId);
 }
