+++ src/main/java/com/backend/api/repositories/ProfileRepository.java
@@ -4,7 +4,7 @@ import com.backend.api.domain.Route;
 import com.backend.api.domain.User;
 import com.backend.api.domain.UserProfile;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
+import com.backend.api.pagination.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.data.jpa.repository.Query;
@@ -15,22 +15,22 @@ import javax.transaction.Transactional;
 import java.util.List;
 
 @Repository(value = "a")
-public interface ProfileRepository extends JpaRepository<UserProfile, Integer>, JpaSpecificationExecutor<UserProfile> {
+public interface ProfileRepository extends JpaRepository<UserProfile, Long>, JpaSpecificationExecutor<UserProfile> {
 
   @Transactional()
   @Query("SELECT x from UserProfile p JOIN p.routes x WHERE p.id = :profileId")
-  Page<Route> getRoutes(@Param("profileId") Integer profileId, Pageable pageable);
+  Page<Route> getRoutes(@Param("profileId") Long profileId, Pageable pageable);
 
   @Transactional()
   @Query("SELECT x from UserProfile p JOIN p.routes x WHERE p.id = :profileId")
-  List<Route> getRoutes(@Param("profileId") Integer profileId);
+  List<Route> getRoutes(@Param("profileId") Long profileId);
 
   @Transactional()
   @Query("SELECT x from UserProfile p JOIN p.users x WHERE p.id = :profileId")
-  Page<User> getUsers(@Param("profileId") Integer id, Pageable pageRequest);
+  Page<User> getUsers(@Param("profileId") Long id, Pageable pageRequest);
 
   @Transactional()
   @Query("SELECT x from UserProfile p JOIN p.users x WHERE p.id = :profileId")
-  List<User> getUsers(@Param("profileId") Integer id);
+  List<User> getUsers(@Param("profileId") Long id);
 
 }
